+++ test/jest.test.ts
@@ -1,4 +1,5 @@
 import { NodeProject, TypeScriptProject } from '../src';
+import { PROJEN_RC } from '../src/common';
 import * as logging from '../src/logging';
 import { mkdtemp, synthSnapshot } from './util';
 
@@ -31,7 +32,6 @@ test('Node Project Jest Defaults Configured', () => {
   const project = new NodeProject({
     outdir: mkdtemp(),
     name: 'test-node-project',
-    start: false,
     mergify: false,
     projenDevDependency: false,
     jest: true,
@@ -54,7 +54,6 @@ test('Node Project Jest With Options Configured', () => {
   const project = new NodeProject({
     outdir: mkdtemp(),
     name: 'test-node-project',
-    start: false,
     mergify: false,
     projenDevDependency: false,
     jest: true,
@@ -81,7 +80,6 @@ test('Typescript Project Jest Defaults Configured', () => {
   const project = new TypeScriptProject({
     outdir: mkdtemp(),
     name: 'test-typescript-project',
-    start: false,
     mergify: false,
     projenDevDependency: false,
     jest: true,
@@ -92,7 +90,7 @@ test('Typescript Project Jest Defaults Configured', () => {
 
   expect(jestTypescriptConfig.compilerOptions).toBeTruthy();
   expect(jestTypescriptConfig.compilerOptions).toStrictEqual(compilerOptionDefaults);
-  expect(jestTypescriptConfig.include).toEqual(['src/**/*.ts', 'test/**/*.ts']);
+  expect(jestTypescriptConfig.include).toEqual([PROJEN_RC, 'src/**/*.ts', 'test/**/*.ts']);
   expect(jestTypescriptConfig.exclude).toEqual(['node_modules']);
 });
 
@@ -106,7 +104,6 @@ test('Typescript Project Jest With Compiler Options', () => {
   const project = new TypeScriptProject({
     outdir: mkdtemp(),
     name: 'test-typescript-project',
-    start: false,
     mergify: false,
     projenDevDependency: false,
     jest: true,